% Generated by roxygen2: do not edit by hand
% Please edit documentation in rdhi.Rmd.
\name{cv.pi}
\alias{cv.pi}
\title{Cross validation for the boost of shortest prediction interval}
\usage{
cv.pi(x, alpha, K = 2)
}
\arguments{
\item{x}{data}

\item{alpha}{nominal error rate}

\item{K}{fold}
}
\description{
Cross validation for the boost of shortest prediction interval
}