% Generated by roxygen2: do not edit by hand
% Please edit documentation in rdhi.Rmd.
\name{data.driven.pi}
\alias{data.driven.pi}
\title{Approximate data driven prediction interval}
\usage{
data.driven.pi(x, alpha)
}
\arguments{
\item{x}{data}

\item{alpha}{nominal error rate}
}
\description{
Approximate data driven prediction interval
}