% Generated by roxygen2: do not edit by hand
% Please edit documentation in rdhi.Rmd.
\name{lower.quantile}
\alias{lower.quantile}
\title{Calculate the lower quantile for data with weights}
\usage{
lower.quantile(v, prob, w = NULL, sorted = FALSE)
}
\arguments{
\item{v}{the data that we want to find quantile}

\item{prob}{find the lower 100*prob\% quantile}

\item{w}{weight for data}

\item{sorted}{whether v has been sorted}
}
\description{
Calculate the lower quantile for data with weights
}