% Generated by roxygen2: do not edit by hand
% Please edit documentation in rdhi.Rmd.
\name{random.pi}
\alias{random.pi}
\title{Calculate shortest prediction interval}
\usage{
random.pi(x, alpha, beta)
}
\arguments{
\item{x}{data}

\item{alpha}{nominal error rate}

\item{beta}{a random position. beta must be less than or equal to alpha.}
}
\description{
Calculate shortest prediction interval
}