% Generated by roxygen2: do not edit by hand
% Please edit documentation in rdhi.Rmd.
\name{shortest.pi}
\alias{shortest.pi}
\title{Calculate shortest prediction interval}
\usage{
shortest.pi(x, alpha)
}
\arguments{
\item{x}{data}

\item{alpha}{nominal error rate}
}
\description{
Calculate shortest prediction interval
}